+++ packages/browser-wallet/src/popup/pages/Account/Earn/Baking/FlowPages/Restake.tsx
@@ -17,7 +17,7 @@ export default function RestakePage({ initial, onNext }: RestakePageProps) {
     const { t } = useTranslation('account', { keyPrefix: 'baking.configure' });
     const { t: tShared } = useTranslation('shared');
     const defaultValues: Partial<RestakePageForm> = useMemo(
-        () => (initial === undefined ? { restake: false } : { restake: initial }),
+        () => (initial === undefined ? { restake: true } : { restake: initial }),
         [initial]
     );
     const onSubmit = (vs: RestakePageForm) => onNext(vs.restake);